requirements.txt
Django == 1.8.2
djangorestframework == 3.4.7
